backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic[email]==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
email-validator==2.1.1
python-socketio==5.10.0
python-engineio==4.8.0
aioredis==2.0.1
